2014-02-28 Roman - !!:SYS: The code clean up for mostly warnings for build by GCC-4.8. - DAQ.AMRDevs: Kontar implementation achives work stage for read properties and clock sync try. - Transport.SSL: Permition to temporary certificate file restrict to 0600.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2058 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/bd/FireBird/firebird.cpp
@@ -146,7 +146,7 @@ void MBD::enable( )
     fdb   = TSYS::strParse(addr(), 0, ";", &off);
     user  = TSYS::strParse(addr(), 0, ";", &off);
     pass  = TSYS::strParse(addr(), 0, ";", &off);
-    conTm = (off < addr().size()) ? TSYS::strParse(addr(), 0, ";", &off) : "1";
+    conTm = (off < (int)addr().size()) ? TSYS::strParse(addr(), 0, ";", &off) : "1";
     cd_pg = codePage().size() ? codePage() : Mess->charset();
 
     //Attach to DB
